+++ public/index.html
@@ -146,6 +146,6 @@
       </section>
     </main>
 
-    <script src="main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
